@@ -8,63 +8,62 @@
     "node_modules"
   ],
   "dependencies": {
-    "@angular/common": "9.1.3",
-    "@angular/compiler": "9.1.3",
-    "@angular/core": "9.1.3",
-    "@angular/forms": "9.1.3",
-    "@angular/localize": "9.1.3",
-    "@angular/platform-browser": "9.1.3",
-    "@angular/platform-browser-dynamic": "9.1.3",
-    "@angular/router": "9.1.3",
+    "@angular/common": "10.0.0",
+    "@angular/compiler": "10.0.0",
+    "@angular/core": "10.0.0",
+    "@angular/forms": "10.0.0",
+    "@angular/localize": "10.0.0",
+    "@angular/platform-browser": "10.0.0",
+    "@angular/platform-browser-dynamic": "10.0.0",
+    "@angular/router": "10.0.0",
     "@fortawesome/angular-fontawesome": "0.6.1",
-    "@fortawesome/fontawesome-svg-core": "1.2.28",
-    "@fortawesome/free-solid-svg-icons": "5.13.0",
+    "@fortawesome/fontawesome-svg-core": "1.2.29",
+    "@fortawesome/free-solid-svg-icons": "5.13.1",
     "@ng-bootstrap/ng-bootstrap": "6.1.0",
     "@ngx-translate/core": "12.1.2",
-    "@ngx-translate/http-loader": "4.0.0",
-    "bootstrap": "4.4.1",
-    "bootswatch": "4.4.1",
-    "moment": "2.24.0",
-    "ng-jhipster": "0.13.0",
+    "@ngx-translate/http-loader": "5.0.0",
+    "bootstrap": "4.5.0",
+    "bootswatch": "4.5.0",
+    "moment": "2.27.0",
+    "ng-jhipster": "0.14.0",
     "ngx-cookie-service": "3.0.4",
-    "ngx-infinite-scroll": "8.0.2",
+    "ngx-infinite-scroll": "9.0.0",
     "ngx-webstorage": "5.0.0",
     "rxjs": "6.5.5",
     "swagger-ui-dist": "3.25.1",
-    "tslib": "1.11.1",
+    "tslib": "2.0.0",
     "zone.js": "0.10.3"
   },
   "devDependencies": {
-    "@angular/cli": "9.1.3",
-    "@angular/compiler-cli": "9.1.3",
-    "@ngtools/webpack": "9.1.3",
-    "@openapitools/openapi-generator-cli": "1.0.12-4.3.0",
-    "@types/jest": "25.2.1",
+    "@angular/cli": "10.0.0",
+    "@angular/compiler-cli": "10.0.0",
+    "@ngtools/webpack": "10.0.0",
+    "@openapitools/openapi-generator-cli": "1.0.13-4.3.1",
+    "@types/jest": "26.0.3",
     "@types/node": "13.13.4",
     "@typescript-eslint/eslint-plugin": "2.30.0",
     "@typescript-eslint/eslint-plugin-tslint": "2.30.0",
     "@typescript-eslint/parser": "2.30.0",
-    "autoprefixer": "9.7.6",
-    "base-href-webpack-plugin": "2.0.0",
+    "autoprefixer": "9.8.4",
     "browser-sync": "2.26.7",
     "browser-sync-webpack-plugin": "2.2.2",
     "codelyzer": "5.2.2",
-    "copy-webpack-plugin": "5.1.1",
-    "css-loader": "3.5.3",
+    "copy-webpack-plugin": "6.0.2",
+    "css-loader": "3.6.0",
     "eslint": "6.8.0",
     "eslint-config-jhipster": "0.0.1",
     "eslint-config-prettier": "6.11.0",
     "eslint-loader": "4.0.2",
     "file-loader": "6.0.0",
     "friendly-errors-webpack-plugin": "1.7.0",
-    "generator-jhipster": "6.9.0",
+    "generator-jhipster": "6.10.1",
     "html-loader": "1.1.0",
-    "html-webpack-plugin": "3.2.0",
+    "html-webpack-plugin": "4.3.0",
     "husky": "4.2.5",
-    "jest": "25.4.0",
+    "jest": "26.1.0",
     "jest-date-mock": "1.0.8",
-    "jest-junit": "10.0.0",
-    "jest-preset-angular": "8.1.3",
+    "jest-junit": "11.0.1",
+    "jest-preset-angular": "8.2.1",
     "jest-sonar-reporter": "2.0.0",
     "lint-staged": "8.2.1",
     "merge-jsons-webpack-plugin": "1.0.21",
@@ -73,22 +72,21 @@
     "optimize-css-assets-webpack-plugin": "5.0.3",
     "postcss-loader": "3.0.0",
     "prettier": "2.0.5",
-    "reflect-metadata": "0.1.13",
     "rimraf": "3.0.2",
-    "sass": "1.26.5",
+    "sass": "1.26.9",
     "sass-loader": "8.0.2",
     "simple-progress-webpack-plugin": "1.1.2",
-    "style-loader": "1.2.0",
-    "terser-webpack-plugin": "2.3.6",
+    "style-loader": "1.2.1",
+    "terser-webpack-plugin": "3.0.6",
     "thread-loader": "2.1.3",
     "to-string-loader": "1.1.6",
-    "ts-loader": "7.0.1",
+    "ts-loader": "7.0.5",
     "tslint": "6.1.2",
-    "typescript": "3.8.3",
+    "typescript": "3.9.5",
     "webpack": "4.43.0",
-    "webpack-bundle-analyzer": "3.7.0",
-    "webpack-cli": "3.3.11",
-    "webpack-dev-server": "3.10.3",
+    "webpack-bundle-analyzer": "3.8.0",
+    "webpack-cli": "3.3.12",
+    "webpack-dev-server": "3.11.0",
     "webpack-merge": "4.2.2",
     "webpack-notifier": "1.8.0",
     "workbox-webpack-plugin": "5.1.3",
