@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpResponse, HttpHeaders } from '@angular/common/http';
 import { DatePipe } from '@angular/common';
-import { ActivatedRoute, ParamMap, Router } from '@angular/router';
+import { ActivatedRoute, ParamMap, Router, Data } from '@angular/router';
+import { combineLatest } from 'rxjs';
 
 import { ITEMS_PER_PAGE } from 'app/shared/constants/pagination.constants';
 import { Audit } from './audit.model';
@@ -13,12 +14,10 @@ import { AuditsService } from './audits.service';
 })
 export class AuditsComponent implements OnInit {
   audits?: Audit[];
-  firstCall = true;
   fromDate = '';
   itemsPerPage = ITEMS_PER_PAGE;
   page!: number;
   predicate!: string;
-  previousPage!: number;
   ascending!: boolean;
   toDate = '';
   totalItems = 0;
@@ -35,44 +34,7 @@ export class AuditsComponent implements OnInit {
   ngOnInit(): void {
     this.toDate = this.today();
     this.fromDate = this.previousMonth();
-    this.activatedRoute.data.subscribe(data => {
-      this.page = data['pagingParams'].page;
-      this.previousPage = data['pagingParams'].page;
-      this.ascending = data['pagingParams'].ascending;
-      this.predicate = data['pagingParams'].predicate;
-      this.loadData();
-    });
-    this.handleBackNavigation();
-  }
-
-  loadPage(page: number): void {
-    this.firstCall = false;
-    this.previousPage = page;
-    this.transition();
-  }
-
-  handleBackNavigation(): void {
-    this.activatedRoute.queryParamMap.subscribe((params: ParamMap) => {
-      if (!this.firstCall) {
-        let prevPage = params.get('page');
-        if (prevPage === null) {
-          prevPage = '1'; // because there are no params in the URL the first time /admin/audits
-        }
-        const prevSort = params.get('sort');
-        const prevSortSplit = prevSort?.split(',');
-        if (prevSortSplit) {
-          this.predicate = prevSortSplit[0];
-          this.ascending = prevSortSplit[1] === 'asc';
-        } else {
-          this.predicate = 'auditEventDate';
-          this.ascending = false;
-        }
-        if (+prevPage !== this.page) {
-          this.page = +prevPage;
-        }
-        this.loadPage(this.page);
-      }
-    });
+    this.handleNavigation();
   }
 
   canLoad(): boolean {
@@ -80,15 +42,15 @@ export class AuditsComponent implements OnInit {
   }
 
   transition(): void {
-    this.firstCall = false;
     if (this.canLoad()) {
       this.router.navigate(['/admin/audits'], {
         queryParams: {
           page: this.page,
           sort: this.predicate + ',' + (this.ascending ? 'asc' : 'desc'),
+          from: this.fromDate,
+          to: this.toDate,
         },
       });
-      this.loadData();
     }
   }
 
@@ -107,6 +69,23 @@ export class AuditsComponent implements OnInit {
     const date = new Date();
     date.setDate(date.getDate() + 1);
     return this.datePipe.transform(date, this.dateFormat)!;
+  }
+
+  private handleNavigation(): void {
+    combineLatest(this.activatedRoute.data, this.activatedRoute.queryParamMap, (data: Data, params: ParamMap) => {
+      const page = params.get('page');
+      this.page = page !== null ? +page : 1;
+      const sort = (params.get('sort') ?? data['defaultSort']).split(',');
+      this.predicate = sort[0];
+      this.ascending = sort[1] === 'asc';
+      if (params.get('from')) {
+        this.fromDate = this.datePipe.transform(params.get('from'), this.dateFormat)!;
+      }
+      if (params.get('to')) {
+        this.toDate = this.datePipe.transform(params.get('to'), this.dateFormat)!;
+      }
+      this.loadData();
+    }).subscribe();
   }
 
   private loadData(): void {
