@@ -32,8 +32,8 @@ import {
   faTrashAlt,
   faAsterisk,
   faTasks,
-  faHome,
   // jhipster-needle-add-icon-import
+  faHome,
 } from '@fortawesome/free-solid-svg-icons';
 
 export const fontAwesomeIcons = [
@@ -70,6 +70,6 @@ export const fontAwesomeIcons = [
   faCalendarAlt,
   faSearch,
   faTrashAlt,
-  faAsterisk,
   // jhipster-needle-add-icon-import
+  faAsterisk,
 ];
