@@ -1,5 +1,6 @@
 import { ComponentFixture, TestBed, async } from '@angular/core/testing';
 import { HttpHeaders, HttpResponse } from '@angular/common/http';
+import { Router, ActivatedRoute } from '@angular/router';
 import { of } from 'rxjs';
 import { advanceTo } from 'jest-date-mock';
 
@@ -8,6 +9,7 @@ import { AuditsComponent } from 'app/admin/audits/audits.component';
 import { AuditsService } from 'app/admin/audits/audits.service';
 import { Audit } from 'app/admin/audits/audit.model';
 import { ITEMS_PER_PAGE } from 'app/shared/constants/pagination.constants';
+import { MockRouter, MockActivatedRoute } from '../../../helpers/mock-route.service';
 
 function build2DigitsDatePart(datePart: number): string {
   return `0${datePart}`.slice(-2);
@@ -36,6 +38,8 @@ describe('Component Tests', () => {
     let comp: AuditsComponent;
     let fixture: ComponentFixture<AuditsComponent>;
     let service: AuditsService;
+    let mockRouter: MockRouter;
+    let mockActivatedRoute: MockActivatedRoute;
 
     beforeEach(async(() => {
       TestBed.configureTestingModule({
@@ -51,6 +55,8 @@ describe('Component Tests', () => {
       fixture = TestBed.createComponent(AuditsComponent);
       comp = fixture.componentInstance;
       service = fixture.debugElement.injector.get(AuditsService);
+      mockRouter = TestBed.get(Router);
+      mockActivatedRoute = TestBed.get(ActivatedRoute);
     });
 
     describe('today function', () => {
@@ -97,7 +103,7 @@ describe('Component Tests', () => {
         expect(comp.toDate).toBe(getDate());
         expect(comp.fromDate).toBe(getDate(false));
         expect(comp.itemsPerPage).toBe(ITEMS_PER_PAGE);
-        expect(comp.page).toBe(10);
+        expect(comp.page).toBe(1);
         expect(comp.ascending).toBe(false);
         expect(comp.predicate).toBe('id');
       });
@@ -121,7 +127,7 @@ describe('Component Tests', () => {
         comp.ngOnInit();
 
         // THEN
-        expect(service.query).toHaveBeenCalled();
+        expect(service.query).toHaveBeenCalledTimes(1);
         expect(comp.audits && comp.audits[0]).toEqual(jasmine.objectContaining(audit));
         expect(comp.totalItems).toBe(1);
       });
@@ -129,18 +135,17 @@ describe('Component Tests', () => {
 
     describe('Create sort object', () => {
       beforeEach(() => {
-        comp.toDate = getDate();
-        comp.fromDate = getDate(false);
         spyOn(service, 'query').and.returnValue(of(new HttpResponse({ body: null })));
       });
 
       it('Should sort only by id asc', () => {
         // GIVEN
-        comp.predicate = 'id';
-        comp.ascending = false;
+        mockActivatedRoute.setParameters({
+          sort: 'id,desc',
+        });
 
         // WHEN
-        comp.transition();
+        comp.ngOnInit();
 
         // THEN
         expect(service.query).toBeCalledWith(
@@ -152,11 +157,12 @@ describe('Component Tests', () => {
 
       it('Should sort by timestamp asc then by id', () => {
         // GIVEN
-        comp.predicate = 'timestamp';
-        comp.ascending = true;
+        mockActivatedRoute.setParameters({
+          sort: 'timestamp,asc',
+        });
 
         // WHEN
-        comp.transition();
+        comp.ngOnInit();
 
         // THEN
         expect(service.query).toBeCalledWith(
@@ -167,36 +173,18 @@ describe('Component Tests', () => {
       });
     });
 
-    describe('loadPage', () => {
-      beforeEach(() => {
-        comp.toDate = getDate();
-        comp.fromDate = getDate(false);
-        comp.previousPage = 1;
-        spyOn(comp, 'transition');
-      });
-
-      it('Should load new page', () => {
-        // WHEN
-        comp.loadPage(2);
-
-        // THEN
-        expect(comp.previousPage).toBe(2);
-        expect(comp.transition).toBeCalled();
-      });
-    });
-
     describe('transition', () => {
-      beforeEach(() => {
-        spyOn(service, 'query').and.returnValue(of(new HttpResponse({ body: null })));
-      });
-
       it('Should not query data if fromDate and toDate are empty', () => {
+        // GIVEN
+        comp.toDate = '';
+        comp.fromDate = '';
+
         // WHEN
         comp.transition();
 
         // THEN
         expect(comp.canLoad()).toBe(false);
-        expect(service.query).not.toBeCalled();
+        expect(mockRouter.navigateSpy).not.toBeCalled();
       });
 
       it('Should query data if fromDate and toDate are not empty', () => {
@@ -209,7 +197,7 @@ describe('Component Tests', () => {
 
         // THEN
         expect(comp.canLoad()).toBe(true);
-        expect(service.query).toBeCalled();
+        expect(mockRouter.navigateSpy).toBeCalled();
       });
     });
   });
